pyproject.toml
[tool.poetry]
name = "peekcell-eln"
version = "0.1.0"
description = ""
authors = ["Louis Nicolle <louisxnicolle@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.4"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.28.0"
django-tables2 = "^2.7.4"
requests = "^2.32.3"
django-bootstrap5 = "^24.3"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
